Use long long instead of double, because double isn't really double, it's same as float, which isn't enough precision. Use MCU registers to access the FAN_ENABLE pin because arduino APIs don't support pin PB6. Make sure variable names always include units (e.g. seconds, millivolts). Cancel low battery alert when charger plugged in. Partially implement reminder when charge is less than 15%. Add new helper class PeriodicCallback. Add renderLongLong. Remove renderDouble(). Fix Timer to properly handle time wraparound.

+++ Product/FanController.cpp
@@ -23,12 +23,12 @@ unsigned int FanController::getRPM() {
 	unsigned long elapsed = millis() - _lastMillis;
 	if (elapsed > _sensorThreshold)
 	{
-		noInterrupts(); // _detachInterrupt();
-		double correctionFactor = 1000.0 / elapsed;
+		noInterrupts();
+		float correctionFactor = 1000.0 / elapsed;
 		_lastReading = correctionFactor * _halfRevs / 2 * 60;
 		_halfRevs = 0;
 		_lastMillis = millis();
-		interrupts(); // _attachInterrupt();
+		interrupts();
 	}
 	return _lastReading;
 }
